Moving performPrefixTransformation into helper file.
