+++ .dev/tests/jest/helpers.js
@@ -1,0 +1,22 @@
+import { getBlockTransforms } from '@wordpress/blocks';
+
+/**
+ * A simplified version of the prefix trigger located in the RichText component.
+ * See: https://github.com/WordPress/gutenberg/blob/53857f67563eb97025d75c196afa643994f0bbcc/packages/block-editor/src/components/rich-text/index.js#L254-L284
+ *
+ * @param {String} blockName The registered block name.
+ * @param {String} prefix The prefix to trigger the transform.
+ * @param {String} content The content of the block when a prefix is found.
+ *
+ * @returns {Object} The block object.
+ */
+export const performPrefixTransformation = ( blockName, prefix, content ) => {
+	const prefixTransforms = getBlockTransforms( 'from', blockName )
+		.filter( ( { type: transformType, prefix: transformPrefix } ) => transformType === 'prefix' && prefix === transformPrefix );
+
+	// Remove prefix trigger from content before performing the transform.
+	const blockContent = content.replace( prefix, '' ).trim();
+	const block = prefixTransforms[ 0 ].transform( blockContent );
+
+	return block;
+};
